requirements.txt

requests
gspread
oauth2client
pyyaml
pytest
